Fixed bugs and completed Remove method. Added exception for unhandled case in remove, fixed bug and changed return type of getServerList in JSONReader.

+++ src/EZShare/Remove.java
@@ -1,9 +1,13 @@
 package EZShare;
 
+import java.net.URI;
+import java.net.URISyntaxException;
+
 import JSON.JSONReader;
 import Resource.HashDatabase;
 import Resource.Resource;
 import exceptions.InvalidResourceException;
+import exceptions.NonExistentResourceException;
 
 /**
  * 
@@ -14,18 +18,36 @@ import exceptions.InvalidResourceException;
  *
  */
 class Remove {
-	static void remove(JSONReader resource, HashDatabase db) throws InvalidResourceException {
+	static void remove(JSONReader resource, HashDatabase db) throws InvalidResourceException, NonExistentResourceException {
 		
 	    String channel = resource.getResourceChannel();
 	    String uri = resource.getResourceUri();
+	    String name = resource.getResourceName();
+	    String desc = resource.getResourceDescription();
+	    String owner = resource.getResourceOwner();
+	    String[] tags = resource.getResourceTags();
+	    
+	    //Check strings etc. are valid.
+	    if(!Common.validateResource(name, desc, tags, uri, channel, owner)){
+	    	throw new InvalidResourceException("Tying to remove resource with invalid fields.");
+	    }
 
-        //valid uri
-	    Common.validUri(uri, "remove");
-
-        //need debug the channel content
-        Resource match = db.uriLookup(channel, uri);
-        if (match != null) {
-            db.deleteResource(match);
+        //validate uri
+	    try {
+            URI path = new URI(uri);
+            if (!path.isAbsolute()) {
+                throw new InvalidResourceException("Trying to remove resource with non-absolute uri.");
+            }
+        } catch (URISyntaxException e) {
+            throw new InvalidResourceException("Attempting to remove resource with invalid uri syntax.");
         }
+	    
+        //Delete according to primary key.
+	    Resource match = db.pKeyLookup(channel, uri);
+	    if(match != null && match.getOwner().equals(owner)){
+	    	db.deleteResource(match);
+	    	return;
+	    }
+	    throw new NonExistentResourceException("Tried to remove a non-existent resource from database.");
 	}
 }
